dom/sdcard/test/test_basic.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=726593
-->
<head>
  <meta charset="utf-8">
  <title>Test for File System API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="filesystem_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for File System API **/

  filesystem_setup();

  var sdcard = navigator.mozSDCard;
  ok(sdcard, "Should have mozSDCard");
  
  var root = sdcard.root;
  ok(root, "Can't find root directory");
  
  var sysName = sdcard.name;
  is(sysName, "SD Card Testing", "Wrong file system name");

  var rootName = sdcard.root.name;
  is(rootName, "file-system-testing", "Wrong root directory name");

  var path = testPath.file.normal.path, name = "testfile";
  root.createFile(path, {ifExists: "fail"}, function(entry) {
	  ok(entry.name == name, "File name should match");

	  root.get(path, function(entry) {
		  ok(entry.name == name, "File name should match");
		  
		  root.remove(path, function() {
			  root.get(path, function(entry) {
				  ok(true, "File was not deleted successfully");
				  filesystem_cleanup();
			  }, function(error) {
				  ok(true, "File was deleted successfully");
				  filesystem_cleanup();
			  });
		  }, function(error) {
			  ok(false, "remove() error: " + error.name);
		  }); 
		  
		  ok(true, "get() success");
	  }, function(error) {
		  ok(false, "get() error: " + error.name);
	  });

	  ok(true, "createFile() success");
  }, function(error) {
	  ok(false, "createFile() error: " + error.name);
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=726593">Mozilla Bug 726593</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
